SEC-2: float ledger → integer smallest-unit (µLENS / µLXC)
Convert every Tier-1 conserved token amount from DOUBLE PRECISION float to
BIGINT integer smallest-unit (1 LENS = 1e6 µLENS; 1 LXC = 1e6 µLXC) so ledger
arithmetic is exact and conservation holds with zero drift. Delete the
roundTo(v,6) band-aid at all call sites + both helper defs; replace the
povi/stakes.go and billing float-epsilon zero-checks with exact integer ==.
House-favoring integer rounding at the 6 meaning-change sites (payouts round
DOWN, charges round UP; remainder retained by the protocol). Emit integer µ-unit
on the JSON boundary with explicit _ulens/_ulxc/_uusd field names. Forward
migrations 0081/0082 ALTER the columns to BIGINT. The SEC-2 drift test flips
RED→GREEN (exact conservation, 0 drift).

STOP before merge — founder review pending. Tier 2/3 (rates/USD) deferred.

+++ cmd/node/rotation_race_test.go
@@ -22,11 +22,11 @@ type countingSlasher struct {
 	n  int
 }
 
-func (s *countingSlasher) Slash(_ context.Context, _ string, fraction float64, _ string) (float64, error) {
+func (s *countingSlasher) Slash(_ context.Context, _ string, fraction float64, _ string) (int64, error) {
 	s.mu.Lock()
 	s.n++
 	s.mu.Unlock()
-	return fraction * 100, nil
+	return int64(fraction * 100), nil
 }
 func (s *countingSlasher) count() int { s.mu.Lock(); defer s.mu.Unlock(); return s.n }
 
@@ -34,7 +34,7 @@ func (s *countingSlasher) count() int { s.mu.Lock(); defer s.mu.Unlock(); return
 type noopChallengeStore struct{}
 
 func (noopChallengeStore) Record(context.Context, povi.Challenge) error { return nil }
-func (noopChallengeStore) UpdateResult(context.Context, string, povi.ChallengeResult, float64, string) error {
+func (noopChallengeStore) UpdateResult(context.Context, string, povi.ChallengeResult, int64, string) error {
 	return nil
 }
 func (noopChallengeStore) AlreadyChallenged(context.Context, string) (bool, error) {
